requirements.txt
numpy
scipy
scikit-learn
scikit-image
smplx==0.1.28
tqdm
yacs
absl-py
ml-collections
tensorboardX
matplotlib
chumpy
opencv-python
torchgeometry
plyfile
pyglet
pymeshlab
pyrender
trimesh
pytorch3d
faiss-gpu